app/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
descope==1.7.9
distro==1.9.0
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.1
Flask==3.1.2
flask-cors==6.0.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-text-splitters==0.3.11
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.25
liccheck==0.9.2
MarkupSafe==3.0.2
openai==1.106.1
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
semantic-version==2.10.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.5.0
zstandard==0.24.0
